implement batch for dual_channel (mitiged results)
